Perf/specialize the ingredient arrays to reduce runtime overhead (#138)

+++ crates/typedown-incremental/src/id/derived.rs
@@ -4,7 +4,7 @@ use super::Id;
 
 /// A fast id for a derived state
 /// Derived id is bound to a database's lifetime
-pub trait DerivedId: Id + From<usize> + Into<usize> {
+pub trait DerivedId: Id + From<u32> + Into<u32> {
   /// Marker used by macros to verify a type implements DerivedId at compile time.
   #[cfg(debug_assertions)]
   #[doc(hidden)]
