+++ crates/typedown-incremental/src/id/input.rs
@@ -2,7 +2,7 @@
 
 /// A fast id for an input state
 /// Input id is bound to a database's lifetime
-pub trait InputId: super::Id + From<usize> + Into<usize> {
+pub trait InputId: super::Id + From<u32> + Into<u32> {
   /// Marker used by macros to verify a type implements InputId at compile time.
   #[cfg(debug_assertions)]
   #[doc(hidden)]
