+++ crates/typedown-incremental/src/id/interned.rs
@@ -1,7 +1,7 @@
 //! Interned id for the incremental database
 
 /// A fast id for an interned state
-pub trait InternedId: super::Id + From<usize> + Into<usize> {
+pub trait InternedId: super::Id + From<u32> + Into<u32> {
   /// Marker used by macros to verify a type implements InternedId at compile time.
   #[cfg(debug_assertions)]
   #[doc(hidden)]
